+++ examples/parsef.rs
@@ -1,4 +1,5 @@
 use std::{
+    env::args,
     io::{Read, stdin},
     process::ExitCode,
 };
@@ -19,11 +20,11 @@ fn main() -> ExitCode {
 fn start() -> Result<()> {
     let mut ip: (u8, u8, u8, u8) = (0, 0, 0, 0);
     let mut mask = 0_u8;
+    let arg = args().nth(1).expect("argument");
 
-    let input: Box<dyn Read> = Box::new(stdin());
     parsef_part!(
-        &mut input.into(),
-        "{}.{}.{}.{}/{mask}\n",
+        &mut arg.into(),
+        "{}.{}.{}.{}/{mask}",
         &mut ip.0,
         &mut ip.1,
         &mut ip.2,