+++ pareg_core/src/check/check_ref.rs
@@ -1,4 +1,4 @@
-use crate::{Reader, Result, SetFromRead};
+use crate::{Reader, Result, SetFromRead, reader::ReadFmt};
 
 /// Wraps [`SetFromRead`] implementation of type so that it also checks for
 /// valid values with the given function.
@@ -11,12 +11,13 @@ pub struct CheckRef<
 impl<T: SetFromRead, F: Fn(&Reader, usize, &T) -> Result<()>> SetFromRead
     for CheckRef<'_, T, F>
 {
-    fn set_from_read(
+    fn set_from_read<'a>(
         &mut self,
         r: &mut Reader,
+        fmt: &'a ReadFmt<'a>,
     ) -> Result<Option<crate::ArgError>> {
-        let pos = r.pos().unwrap_or_default();
-        match self.0.set_from_read(r) {
+        let pos = r.pos();
+        match self.0.set_from_read(r, fmt) {
             Ok(res) => self.1(r, pos, self.0).map(|_| res),
             e => e,
         }