+++ pareg_core/src/parsef.rs
@@ -1,22 +1,22 @@
 use std::borrow::Cow;
 
-use crate::{ArgError, Reader, Result, SetFromRead};
+use crate::{ArgError, Reader, Result, SetFromRead, reader::ReadFmt};
 
 /// Argument to [`parsef`] describing expected operation.
-pub enum ParseFArg<'a> {
+pub enum ParseFArg<'a, 'f> {
     /// Expect a string.
     Str(Cow<'a, str>),
-    /// Expect to parse to the given value.
-    Arg(&'a mut dyn SetFromRead),
+    /// Expect to parse to the given value with the given format.
+    Arg(&'a mut dyn SetFromRead, &'f ReadFmt<'f>),
 }
 
 /// Parsef implementation. Parse all data in `r` based on `args`.
 ///
 /// This is usually used by the `parsef!` macro, but nothing forbids you from
 /// constructing the parse operation at runtime.
-pub fn parsef<'a>(
+pub fn parsef<'a, 'f>(
     r: &mut Reader,
-    args: impl AsMut<[ParseFArg<'a>]>,
+    args: impl AsMut<[ParseFArg<'a, 'f>]>,
 ) -> Result<()> {
     let res = parsef_part(r, args)?;
     if r.peek()?.is_none() {
@@ -30,14 +30,14 @@ pub fn parsef<'a>(
 ///
 /// This is usually used by the `parsef_part!` macro, but nothing forbids you
 /// from constructing the parse operation at runtime.
-pub fn parsef_part<'a>(
+pub fn parsef_part<'a, 'f>(
     r: &mut Reader,
-    mut args: impl AsMut<[ParseFArg<'a>]>,
+    mut args: impl AsMut<[ParseFArg<'a, 'f>]>,
 ) -> Result<Option<ArgError>> {
     let mut last_err = None;
     for a in args.as_mut() {
         last_err = match a {
-            ParseFArg::Arg(a) => a.set_from_read(r)?,
+            ParseFArg::Arg(a, fmt) => a.set_from_read(r, fmt)?,
             ParseFArg::Str(a) => {
                 r.expect(a)?;
                 None