+++ tests/check.rs
@@ -16,12 +16,14 @@ fn test_check_ref() {
     };
 
     assert!(matches!(
-        CheckRef(&mut n, &check_fun).set_from_read(&mut "8".into()),
+        CheckRef(&mut n, &check_fun)
+            .set_from_read(&mut "8".into(), &"".into()),
         Ok(None)
     ));
     assert_eq!(n, 8);
     assert!(matches!(
-        CheckRef(&mut n, &check_fun).set_from_read(&mut "9".into()),
+        CheckRef(&mut n, &check_fun)
+            .set_from_read(&mut "9".into(), &"".into()),
         Err(_)
     ));
     assert_eq!(n, 9);
@@ -56,27 +58,32 @@ fn test_in_range() {
     let mut n = 0_i32;
 
     assert!(matches!(
-        check::InRange(&mut n, 0..100).set_from_read(&mut "20".into()),
+        check::InRange(&mut n, 0..100)
+            .set_from_read(&mut "20".into(), &"".into()),
         Ok(_)
     ));
     assert_eq!(n, 20);
     assert!(matches!(
-        check::InRange(&mut n, 0..100).set_from_read(&mut "0".into()),
+        check::InRange(&mut n, 0..100)
+            .set_from_read(&mut "0".into(), &"".into()),
         Ok(_)
     ));
     assert_eq!(n, 0);
     assert!(matches!(
-        check::InRange(&mut n, 0..100).set_from_read(&mut "99".into()),
+        check::InRange(&mut n, 0..100)
+            .set_from_read(&mut "99".into(), &"".into()),
         Ok(_)
     ));
     assert_eq!(n, 99);
     assert!(matches!(
-        check::InRange(&mut n, 0..100).set_from_read(&mut "-1".into()),
+        check::InRange(&mut n, 0..100)
+            .set_from_read(&mut "-1".into(), &"".into()),
         Err(_)
     ));
     assert_eq!(n, -1);
     assert!(matches!(
-        check::InRange(&mut n, 0..100).set_from_read(&mut "100".into()),
+        check::InRange(&mut n, 0..100)
+            .set_from_read(&mut "100".into(), &"".into()),
         Err(_)
     ));
     assert_eq!(n, 100);